refactor(node): split Node class into functional modules
